+++ packages/app/src/pages/happiness-training/index.tsx
@@ -24,8 +24,8 @@ import Loading from "components/loading";
 import { useIdentity } from "context/identity";
 import { NetworkStatus } from "apollo-client";
 // import { doScheduleNotification } from "../../context/happiness/notification";
-import { useData } from "./useData";
 import Slide, { styles as slidesStyles } from "./Slide";
+import { useData } from "./use-data";
 import { slideGutter, slideWidth } from "./constants";
 import { IconSvg } from "../../components/icon";
 import Colors from "../../colors";
